abstract class controller and new controller classes
